declare types in correct modules
